requirements.txt
asgiref==3.5.2
Django==4.1.3
django-appconf==1.0.5
django-compressor==4.1
django-tinymce==3.5.0
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.3
